Disable hardware-accelerated thumbnailing mode (WIP)
Thumbnailing mode appears to be causing crashes with the avc codec.

Thumbnailing mode is triggered when indexing files in the Gallery app,
but it also appears that the camera uses the
OMX.qcom.video.decoder.avc codec in thumbnail mode when in
viewfinder/snapshot mode with certain cameraHAL versions (a bug)?

For an example of the camera crashing without this patch (Galaxy 5):
https://gist.github.com/4444260

This change fixes the camera & thumbnailing crashes, but the
root cause needs to be identified/fixed properly.

Change-Id: I9312e609e45d218163a8ee2c415a0a126bdbf3fe